fix: adapt lexer to grammar defined in the project file
